+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   or a preview image if the client supports it ([#1261](https://github.com/latex-lsp/texlab/issues/1261))
 - Add `texlab.symbols.customEnvironments` setting for specifying additional environments that will be included in the document symbols
   ([#1292](https://github.com/latex-lsp/texlab/issues/1292))
+- Add `texlab.experimental.labelReferenceRangeCommands` setting ([#1210](https://github.com/latex-lsp/texlab/issues/1210))
 
 ## [5.21.0] - 2024-10-26